+++ src/samples/create-payout.ts
@@ -1,0 +1,47 @@
+/**
+ *  Copyright 2018 Angus.Fenying <fenying@litert.org>
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+// tslint:disable:no-console
+import * as Paypal from "../libs";
+
+(async () => {
+
+    const cli: Paypal.IClient = Paypal.createClient(
+        require(`${__dirname}/../config.json`)
+    );
+
+    let cmd = new Paypal.APIs.v1.Payouts.CreatePayout.Command(
+        process.argv[2],
+        process.argv[3],
+        process.argv[4],
+        [{
+            recipient_type: "EMAIL",
+            receiver: process.argv[5],
+            amount: {
+                currency: process.argv[6],
+                value: process.argv[7]
+            },
+            sender_item_id: process.argv[8],
+            note: process.argv[9]
+        }]
+    );
+
+    console.log(JSON.stringify(await cli.execute(cmd), null, 2));
+
+})().catch((e) => {
+
+    console.error(e);
+});