+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes Logs
 
+## v0.2.0
+
+- Add APIs `v1.Payouts.CreatePayout`.
+- Add APIs `v1.Payouts.ViewPayout`.
+
 ## v0.1.1
 
 - Fixed: Wrong position for `related_resources`.